example05とexample06のtypoも修正
